+++ change-log.md
@@ -4,8 +4,9 @@
 ## Genshin Impact v. 6.6 pt. 2
 * Added Lohen, 5* Cryo polearm
 
-
 ## Updates
+* Updated update system to hopefully make it obsolete to manually have to install from the Github page
+* Added in-app update status toasts for download progress and install readiness
 * Linked Lohen to his signature weapon
 * Added Pneuma and Ousia bonus abilities